created new role view in admin dashboard; created RoleDelail.view; upd role CRUD; add chenged permissions in role; upd vue-routes and api.php; created RoleEditName,RoleUpdatePermissions request; created RoleUserResource; delete EditRoleModal.vue;

+++ app/Http/Controllers/Admin/Role/RoleController.php
@@ -3,32 +3,52 @@
 namespace App\Http\Controllers\Admin\Role;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\Admin\Role\RoleEditNameRequest;
 use App\Http\Requests\Admin\Role\RoleRequest;
+use App\Http\Requests\Admin\Role\RoleUpdatePermissionsRequest;
 use App\Http\Resources\Admin\Role\RoleResource;
 use App\Models\Role;
 
 class RoleController extends Controller
 {
 
-    public function index()
+    /**
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function index(): \Illuminate\Http\JsonResponse
     {
         return response()->json(['roles' => RoleResource::collection(Role::all())]);
     }
 
-    public function store(RoleRequest $request)
+    /**
+     * @param RoleRequest $request
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function store(RoleRequest $request): \Illuminate\Http\JsonResponse
     {
         $data = $request->validated();
         $data['slug'] = str_replace(' ', '-', strtolower($data['name']));
+        $permissions = $data['permissions']; unset($data['permissions']);
         $role = Role::firstOrCreate($data);
+        $role->permissions()->attach($permissions);
         return response()->json(['message' => 'Role created!']);
     }
 
+    /**
+     * @param Role $role
+     * @return object|\Illuminate\Http\JsonResponse
+     */
     public function show(Role $role): object
     {
         return response()->json(['role' => new RoleResource($role)]);
     }
 
-    public function update(RoleRequest $request, Role $role)
+    /**
+     * @param RoleEditNameRequest $request
+     * @param Role $role
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function update(RoleEditNameRequest $request, Role $role): \Illuminate\Http\JsonResponse
     {
         $data = $request->validated();
         $data['slug'] = str_replace(' ', '-', strtolower($data['name']));
@@ -39,16 +59,40 @@ class RoleController extends Controller
         return response()->json(['message' => 'Role updated!']);
     }
 
-    public function delete(Role $role)
+    /**
+     * @param Role $role
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function delete(Role $role): \Illuminate\Http\JsonResponse
     {
+        if($role->users->count()){
+            return response()->json(['message' => 'You cannot delete this role because there are users associated with it. Assign users a different role.']);
+        }
         $role->delete();
         return response()->json(['message' => 'Role deleted successfully!']);
     }
 
-    public function status(Role $role)
+    /**
+     * @param Role $role
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function status(Role $role): \Illuminate\Http\JsonResponse
     {
         $role->status = !$role->status;
         $role->save();
         return response()->json(['message' => 'Role status changed!']);
+    }
+
+    /**
+     * @param RoleUpdatePermissionsRequest $request
+     * @param Role $role
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function changePermissions(RoleUpdatePermissionsRequest $request, Role $role): \Illuminate\Http\JsonResponse
+    {
+        $data = $request->validated();
+        $role->permissions()->sync($data['permissions']);
+        $role->save();
+        return response()->json(['message' => 'Role permission change successfully!']);
     }
 }
